DocuMind/requirements.txt
pypdf
faiss-cpu
sentence-transformers
numpy
tqdm
